Chỉnh sửa các entity cho đúng với sơ đồ lớp

+++ src/main/java/iuh/fit/son23641341/nhahanglau_phantan/dao/KhachHang_DAO.java
@@ -102,7 +102,7 @@ public class KhachHang_DAO {
         return MockData.khachHangs().removeIf(item -> item.getMaKhachHang().equals(maKhachHang));
     }
 
-    public boolean congDiemTichLuy(String maKhachHang, float diemCong) {
+    public boolean congDiemTichLuy(String maKhachHang, int diemCong) {
         KhachHang existing = MockData.khachHangs().stream()
             .filter(item -> item.getMaKhachHang().equals(maKhachHang))
             .findFirst()
